Gather version handling code
